PDF Genertor/requirements.txt
Pillow==11.1.0
reportlab==4.3.1
